Split MainAppView into smaller files under the 'Views' folder and changed the survey button to be liquid glass with different color indicators
